// frontend/assessments/3/src/types.ts
// Generated by https://quicktype.io

export interface RoomType {
    id:           number;
    name:         string;
    costPerNight: string;
    currency:     string;
    addOns:       AddOn[];
}

export interface AddOn {
    name:     string;
    cost:     string;
    currency: string;
}


export interface RoomTypeSlice {
    rooms : RoomType[];
    loading : boolean;
    error : string;
}

export type Room = 'None' | "Single Room" | "Twin Room" | "Deluxe" | "Presidential Suite";